l2eb_data: make UPD optional
The Saturn version uses the same packed data format as the PlayStation version,
with file `2` serving the role of `DATA.PAK` and `1` serving the role of
`DATA.IDX`. However, the Saturn has no filename mapping table; the files are
all unnamed.

This commit updates l2eb_data to allow unpacking a file without an `.UPD` file;
files will just be given numeric filenames in this case.

In order to accommodate varying numbers of arguments, the .UPD file and the
filename prefix have been turned into options.